refactor: haversine 等の共通ロジックを shared/geoUtils.ts に抽出
distance デモと plan デモで haversineDistanceMeters /
formatHorizontalDistance / formatAltitudeDelta が重複実装されていた。
src/demos/shared/geoUtils.ts に共通実装を移動し、両デモから re-export
する形に変更した。

これにより haversine ロジックの修正は 1 か所で済み、挙動のズレを防げる。
